Bug 1674777 part 2 - Enable large ArrayBuffers for jsapi-tests. r=sfink
Depends on D103273

Differential Revision: https://phabricator.services.mozilla.com/D103274

+++ js/src/jsapi-tests/testTypedArrays.cpp
@@ -116,7 +116,7 @@ template <JSObject* Create(JSContext*, size_t), typename Element,
                            const JS::AutoRequireNoGC&)>
 bool TestPlainTypedArray(JSContext* cx) {
   {
-    RootedObject notArray(cx, Create(cx, UINT32_MAX));
+    RootedObject notArray(cx, Create(cx, SIZE_MAX));
     CHECK(!notArray);
   }
